modification done to run application

+++ backend/FleetEase-Backend/src/main/java/com/fleetease/backend/repository/CarMaintainanceRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.fleetease.backend.model.Car;
 import com.fleetease.backend.model.CarMaintainance;
 
-public interface CarMaintainanceRepo extends JpaRepository<CarMaintainance, Integer> {
+public interface CarMaintainanceRepo extends JpaRepository<CarMaintainance, Long> {
 
 	List<CarMaintainance> findByCar(Car car);
 }
